Implement a ping command

+++ TestC21/src/can_utils.h
@@ -11,7 +11,7 @@ typedef struct _can_hw_filter {
 
 int can_filter_add(CAN_HW_FILTER *pfilter);
 int can_filter_switch(uint8_t on, uint8_t index, uint8_t ext);
-int can_filter_remove(uint8_t index, uint8_t ext);
+int can_filter_remove(int index, uint8_t ext);
 void can_filters_init(void);
 int can_msg_get(int timeout);
 int can_msg_free(int index);
